refactor(invite-member-modal): state fixes; change button to invite-button; rename component to invite-modal... (from add-modal...); and some else...
