--- app/fragments/app-add-member-modal/src/index.ts
@@ -1,1 +1,0 @@
-export * from './add-member-modal.component.js'
