+++ app/fragments/app-invite-member-modal/src/hooks/change-button.interface.ts
@@ -1,0 +1,7 @@
+import { InviteButtonStateType } from '@ui/invite-button'
+
+export interface ChangeButtonHookType {
+  inputValues: Array<string>
+  checkedSwitches: Array<string>
+  setInviteButtonState: (buttonState: InviteButtonStateType) => void
+}
