+++ app/fragments/app-invite-member-modal/src/index.ts
@@ -1,0 +1,1 @@
+export * from './invite-member-modal.component.js'
