+++ app/fragments/app-invite-member-modal/src/input/input.constants.ts
@@ -1,0 +1,3 @@
+export const INPUT_PROPS = {
+  placeholder: 'me@torinasakura.name',
+}
