+++ app/fragments/app-invite-member-modal/src/invite-member-modal.interfaces.ts
@@ -1,4 +1,4 @@
-export interface AddMemberModalProps {
+export interface InviteMemberModalProps {
   open: boolean
   onBackdropClick: VoidFunction
 }
