+++ app/fragments/app-invite-member-modal/src/ivnite-member-modal.constants.ts
@@ -1,0 +1,7 @@
+export const getIconProps = (theme) => {
+  return {
+    width: theme.spaces.large,
+    height: theme.spaces.large,
+    color: 'none',
+  }
+}
